# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A REST API for managing spy cats, missions, and agents",
        "title": "Spy Cat Agency API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/v1",
    "paths": {
...
        "/spy-cats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all spy cats",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spy-cats"
                ],
                "summary": "List all spy cats",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SpyCatsResponseDoc"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponseDoc"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new spy cat with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spy-cats"
                ],
                "summary": "Create a new spy cat",
                "parameters": [
                    {
                        "description": "Spy Cat Details",
                        "name": "spy-cat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CreateSpyCatRequestDoc"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.SpyCatResponseDoc"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponseDoc"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponseDoc"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/main.ValidationErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponseDoc"
                        }
                    }
                }
            }
        },
        "/spy-cats/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get details of a specific spy cat by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spy-cats"
                ],
                "summary": "Get a spy cat by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Spy Cat ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SpyCatResponseDoc"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponseDoc"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponseDoc"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a spy cat by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spy-cats"
                ],
                "summary": "Delete a spy cat",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Spy Cat ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.MessageResponseDoc"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponseDoc"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponseDoc"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the salary of a specific spy cat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spy-cats"
                ],
                "summary": "Update spy cat salary",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Spy Cat ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Salary Update",
                        "name": "salary",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.UpdateSpyCatSalaryRequestDoc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SpyCatResponseDoc"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponseDoc"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponseDoc"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponseDoc"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.CreateSpyCatRequestDoc": {
            "description": "Request body for creating a new spy cat",
            "type": "object",
            "properties": {
                "breed": {
                    "description": "Cat breed\nExample: Siamese",
                    "type": "string"
                },
                "name": {
                    "description": "Spy cat name\nExample: Agent Whiskers",
                    "type": "string"
                },
                "password": {
                    "description": "Password for authentication\nExample: secretpassword123",
                    "type": "string"
                },
                "salary": {
                    "description": "Annual salary\nExample: 50000.00",
                    "type": "number"
                },
                "years_of_experience": {
                    "description": "Years of experience\nExample: 5",
                    "type": "integer"
                }
            }
        },
...
        "main.ErrorResponseDoc": {
            "description": "Standard error response format",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message\nExample: error message",
                    "type": "string"
                }
            }
        },
        "main.MessageResponseDoc": {
            "description": "Simple message response format",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Success message\nExample: operation completed successfully",
                    "type": "string"
                }
            }
        },
...
        "main.SpyCatDoc": {
            "description": "Spy cat entity",
            "type": "object",
            "properties": {
                "breed": {
                    "description": "Cat breed\nExample: Siamese",
                    "type": "string"
                },
                "id": {
                    "description": "Unique identifier\nExample: 1",
                    "type": "integer"
                },
                "name": {
                    "description": "Spy cat name\nExample: Agent Whiskers",
                    "type": "string"
                },
                "salary": {
                    "description": "Annual salary\nExample: 50000.00",
                    "type": "number"
                },
                "years_of_experience": {
                    "description": "Years of experience\nExample: 5",
                    "type": "integer"
                }
            }
        },
        "main.SpyCatResponseDoc": {
            "description": "Response containing a single spy cat",
            "type": "object",
            "properties": {
                "spy-cat": {
                    "description": "Spy cat data",
                    "allOf": [
                        {
                            "$ref": "#/definitions/main.SpyCatDoc"
                        }
                    ]
                }
            }
        },
        "main.SpyCatsResponseDoc": {
            "description": "Response containing a list of spy cats",
            "type": "object",
            "properties": {
                "spy-cats": {
                    "description": "List of spy cats",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.SpyCatDoc"
                    }
                }
            }
        },
...
        "main.UpdateSpyCatSalaryRequestDoc": {
            "description": "Request body for updating a spy cat's salary",
            "type": "object",
            "properties": {
                "salary": {
                    "description": "New salary amount\nExample: 55000.00",
                    "type": "number"
                }
            }
        },
...
        "main.ValidationErrorResponseDoc": {
            "description": "Validation error response format with field-specific errors",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Map of field names to error messages\nExample: {\"name\": \"must be provided\", \"breed\": \"invalid breed\"}",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}